Move SpanFactory into internal package. No need to be in the public API.
