+++ api/src/main/java/io/opencensus/trace/SpanBuilder.java
@@ -15,6 +15,7 @@ package io.opencensus.trace;
 
 import io.opencensus.common.NonThrowingCloseable;
 import io.opencensus.trace.base.EndSpanOptions;
+import io.opencensus.trace.internal.SpanFactory;
 import java.util.List;
 import javax.annotation.Nullable;
