+++ api/src/main/java/io/opencensus/trace/internal/SpanFactory.java
@@ -11,8 +11,11 @@
  * limitations under the License.
  */
 
-package io.opencensus.trace;
+package io.opencensus.trace.internal;
 
+import io.opencensus.trace.Span;
+import io.opencensus.trace.SpanContext;
+import io.opencensus.trace.StartSpanOptions;
 import javax.annotation.Nullable;
 
 /** Factory class to create and start a {@link Span}. */
@@ -26,7 +29,7 @@ public abstract class SpanFactory {
    * @param options The options for the start of the {@code Span}.
    * @return A child {@code Span} that will have the name provided.
    */
-  protected abstract Span startSpan(@Nullable Span parent, String name, StartSpanOptions options);
+  public abstract Span startSpan(@Nullable Span parent, String name, StartSpanOptions options);
 
   /**
    * Creates and starts a new child {@link Span} (or root if parent is {@code null}), with parent
@@ -39,6 +42,6 @@ public abstract class SpanFactory {
    * @param options The options for the start of the {@code Span}.
    * @return A child {@code Span} that will have the name provided.
    */
-  protected abstract Span startSpanWithRemoteParent(
+  public abstract Span startSpanWithRemoteParent(
       @Nullable SpanContext remoteParent, String name, StartSpanOptions options);
 }
