+++ api/src/test/java/io/opencensus/trace/TracerTest.java
@@ -26,6 +26,7 @@ import io.opencensus.trace.base.EndSpanOptions;
 import io.opencensus.trace.base.SpanId;
 import io.opencensus.trace.base.TraceId;
 import io.opencensus.trace.base.TraceOptions;
+import io.opencensus.trace.internal.SpanFactory;
 import java.util.Random;
 import org.junit.Before;
 import org.junit.Rule;
